Fixing the iOS ( build
* Fixed some compile errors on iOS.
* Updated xcconfig files for tests on iOS.
* Copy test data files into test bundle on iOS, and set up
  SFixturesDir accordingly.
* Configured "LiteCore XCTests" as the test bundle for the
  "LiteCore-static" target and created a scheme for it. If you select
  the "LiteCore XCTests" scheme, select a simulator destination,
  and then Test (Cmd-U) it will run the unit tests in the simulator.
* Running tests on iOS devices TBD…

+++ Networking/TCPSocket.cc
@@ -454,7 +454,7 @@ namespace litecore { namespace net {
         if (contentLength >= 0) {
             // Read exactly Content-Length bytes:
             if (contentLength > 0) {
-                body.resize(contentLength);
+                body.resize(size_t(contentLength));
                 if (readExactly((void*)body.buf, (size_t)contentLength) < contentLength) {
                     body.reset();
                     return false;
